Add CLI option for running just pre-fill
Adds a command line option `apPreFill` that runs the randomizer process up until the point where the items are actually shuffled, and produces an output file with the logic requirements for every check and event flag, as well as a list of item names with relevant meta-data.  Currently, the intent is for the Archipelago generation code to call the CLI with this option, then parse the resulting file in order to create items, locations, and logic rules. This is mostly a proof of concept; formatting not final.