Move AP communication bytes into save data
There was previously a small chance that BizHawkClient could send an item to the game and update the received item index, then the game would save before actually getting the item in-game, saving the updated index but not the item itself, causing the item to get permanently lost on a reset. Moving the communication bytes into the save data means now in that situation the item will be immediately received upon reloading the save file.